moving state management 1 level below
