+++ client/src/components/Checkout/BookingForm/BookingForm.jsx
@@ -5,20 +5,114 @@ import PhoneInput from 'react-phone-number-input';
 import 'react-phone-number-input/style.css';
 import Customer from './Customer/Customer';
 import Registration from './RegistrationForm/Registration';
+import { observer, inject } from "mobx-react";
+// inject("colors")(observer(({ colors, label, onClick })
+
+// const BookingForm = inject("customerStore", "checkoutFormStore")(observer(({
+//   customerStore, checkoutFormStore, checkoutFormValidations, onCancel, phoneChanged,
+//   foundCustomer, newCustomerHandler, onRegistrationSubmit, onCheckoutInputSubmit, registrationFormValidations, carCheckoutHandler, phone,
+// }) => {
+@inject("customerStore", "individualCarStore", "checkoutFormStore")
+@observer
+export default class BookingForm extends React.Component {
+  constructor(props) {
+    super(props);
+    this.customerStore = this.props.customerStore;
+    this.individualCarStore = this.props.individualCarStore;
+    this.checkoutFormStore = this.props.checkoutFormStore;
+  }
 
 
-const BookingForm = ({
-  changed, checkoutFormValidations, onCancel, phoneChanged,
-  foundCustomer, newCustomerHandler, onRegistrationSubmit, onCheckoutInputSubmit, registrationFormValidations, carCheckoutHandler, phone,
-}) => {
+  handlePhoneChanged = (value) => {
+    let phoneNumber = "";
+    const { phone } = this.customerStore;
+    if (value !== undefined) {
+      phoneNumber = parsePhoneNumber(value);
+    }
+
+    const newObj = {};
+    newObj.touched = true;
+    newObj.value = value;
+
+    if (phoneNumber) {
+      if (phoneNumber.country === "BG") {
+        if (newObj.value.length === 13) {
+          newObj.isValid = true;
+        }
+      }
+    }
+
+    this.customerStore.phone = newObj;
+
+    if (newObj.isValid) {
+      fetchRequestCustomer(`${baseURL}/${customers}`, "PUT", {
+        phone: value,
+      }).then((response) => {
+        if (response) {
+          this.customerStore.foundCustomer = response;
+        }
+      });
+    } else {
+      this.customerStore.foundCustomer = {
+        phone: "",
+        firstName: "",
+        lastName: "",
+        birthdate: "",
+        age: "",
+        loyaltyDiscount: "",
+        geoDiscount: "",
+      };
+    }
+  };
+
+  newCustomerHandler = (event) => {
+    const name = event.target.dataset.name;
+    const value = event.target.value;
+    const newObj = this.customerStore.newCustomer;
+    const { registrationFormValidations } = this.customerStore;
+    const validationObj = registrationFormValidations;
+    newObj[name] = value;
+    newObj.phone = this.customerStore.phone.value;
+    validationObj[name].valid = checkInputValidity(
+      newObj[name],
+      registrationFormValidations[name].rules
+    );
+    validationObj[name].touched = true;
+
+    this.customerStore.newCustomer = Object.assign(this.customerStore.newCustomer, newObj);
+    this.customerStore.registrationFormValidations = Object.assign(this.customerStore.registrationFormValidations, validationObj);
+  };
+
+  onRegistrationSubmit = (event) => {
+    const { newCustomer, registrationFormValidations } = this.customerStore;
+    const age = differenceInYears(newCustomer.birthdate);
+    bookingFormErrors(newCustomer, registrationFormValidations);
+    if (
+      !registrationFormValidations.firstName.valid ||
+      !registrationFormValidations.lastName.valid ||
+      !(age >= 18) ||
+      !registrationFormValidations.birthdate.touched ||
+      !registrationFormValidations.birthdate.valid
+    ) {
+      return;
+    }
+    fetchRequest(`${baseURL}/${customers}`, "POST", newCustomer).then(
+      (response) => {
+        this.customerStore.foundCustomer = response;
+      }
+    );
+  };
+
+  render() {
   let names = <div>hi</div>;
+  const { foundCustomer, phone, newCustomer, registrationFormValidations } = this.customerStore;
 
   if (foundCustomer.firstName === '') {
     names = (
       <Registration
-        newCustomerHandler={newCustomerHandler}
+        newCustomerHandler={this.newCustomerHandler}
         registrationFormValidations={registrationFormValidations}
-        onRegistrationSubmit={onRegistrationSubmit}
+        onRegistrationSubmit={this.onRegistrationSubmit}
       />
     );
   } else {
@@ -63,7 +157,7 @@ const BookingForm = ({
               placeholder="Enter phone number"
               data-name="phone"
               required
-              onChange={phoneChanged}
+              onChange={this.phoneChanged}
             />
             {found}
           </div>
@@ -74,7 +168,7 @@ const BookingForm = ({
 
     </div>
   );
-};
+}};
 
 BookingForm.propTypes = {
   changed: PropTypes.func,
@@ -106,4 +200,3 @@ BookingForm.defaultProps = {
 };
 
 
-export default BookingForm;
